DEV-010 Custom Exception

+++ app/Application/UseCase/ProductDeleteUseCase.php
@@ -3,6 +3,7 @@
 namespace App\Application\UseCase;
 
 use App\Application\UseCase\Contract\IProductDeleteUseCase;
+use App\Exceptions\NotFoundException;
 use App\Infrastructure\Product\Contract\IProductDeleteRepository;
 
 class ProductDeleteUseCase implements IProductDeleteUseCase
@@ -16,6 +17,10 @@ class ProductDeleteUseCase implements IProductDeleteUseCase
 
     public function execute(int $id)
     {
+        if (! is_numeric($id) || $id <= 0) {
+            throw new NotFoundException('Invalid id', 400);
+        }
+
         return $this->productRepository->delete($id);
     }
 }
